slm-api/requirements.txt
fastapi
uvicorn[standard]
transformers>=4.35.0
torch
accelerate
python-dotenv
